fix: VS2019 でエラーになるコードを修正。

+++ Scratch/LowLevelGraphics/VulkanSample/AeVulkanLib/ae/base/EnumKeyArray.hpp
@@ -85,7 +85,7 @@ public:
     }
     //@}
 
-    ValueType values_[TEnumType::TERM];
+    ValueType values_[static_cast<int>(TEnumType::TERM)];
 };
 
 } // namespace base
